requirements-rt.txt
fastapi>=0.135.3
uvicorn[standard]>=0.44.0
redis>=7.4.0
python-dotenv>=1.2.2
pydantic>=2.0
